requirements.txt
flask
transformers
torch
pyspellchecker
flask-cors
